Remove Insights from namespace
