+++ Dockerfile
@@ -14,7 +14,7 @@ RUN yum -y install centos-release-scl-rh && \
                    && \
     yum clean all
 
-ENV WORKDIR /opt/insights-topological_inventory/
+ENV WORKDIR /opt/topological_inventory/
 ENV RAILS_ROOT $WORKDIR
 WORKDIR $WORKDIR
