+++ config/initializers/session_store.rb
@@ -1,3 +1,3 @@
 # Be sure to restart your server when you modify this file.
 
-Rails.application.config.session_store :cookie_store, key: '_insights-topological_inventory_session'
+Rails.application.config.session_store :cookie_store, key: '_topological_inventory_session'
