+++ docker-assets/start_persister
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-bundle exec insights-topological_inventory-persister
+bundle exec topological_inventory-persister
