--- lib/insights/topological_inventory/schema.rb
@@ -1,8 +1,0 @@
-require "insights/topological_inventory/schema/default"
-
-module Insights
-  module TopologicalInventory
-    module Schema
-    end
-  end
-end
