+++ lib/topological_inventory/schema.rb
@@ -1,0 +1,6 @@
+require "topological_inventory/schema/default"
+
+module TopologicalInventory
+  module Schema
+  end
+end
